==> Codes/0_6_merge_datasets.R <==
## -------------------------------------------------------------------
base_counties <- readRDS(here("Data/data-processed/base_counties.rds"))

#' Weather
weather_data <- readRDS(here("Data/data-processed/summarized_gridMET.rds"))

#' Saturated thickness
sat_all <- readRDS(here("Data/data-processed/sat_all.rds"))

#' Yield, acreage data
nass_data <- readRDS(here("Data/data-processed/nass_data.rds"))

#* SSURGO
ssurgo_soil <- readRDS(here("Data/data-processed/ssurgo_soil.rds"))


## -------------------------------------------------------------------
merged_data <-
  base_counties[nass_data, on = c("sc_code")] %>%
  weather_data[., on = c("sc_code", "year")] %>%
  sat_all[., on = c("sc_code", "year")] %>%
  .[order(sc_code, year), ] %>%
  ssurgo_soil[, on = "sc_code"]


## -------------------------------------------------------------------
final_data <-
  merged_data %>%
  #--- yield unit conversion (bu/acre to ton/ha)---#
  .[, yield := yield * 12.553 / 200] %>%
  # === if not in_hpa and irrigated, drop them ===#
  # irrigated, but saturated thickness not observed. These
  # observations cannot be used.
  .[!(in_hpa == 0 & ir == "ir"), ] %>%
  #--- drop if irrigated and year > 2016 ---#
  # drops irrigated produciton observations that are missing sat (2017, 2018 fow which sat is not observed). This will keep all the dryland production observations beyond 2016 because sat does not have to be observed for the dryland production.
  .[!(ir == "ir" & year > 2016), ] %>%
  # === drop if both yield and acres are not observed ===#
  .[!is.na(yield) & !is.na(acres), ] %>%
  #--- sc_code and irrigation status (used for FE regression) ---#
  .[, sc_dry := paste0(sc_code, "_", ir)] %>%
  #--- state-year combination ---#
  .[, state_year := paste0(state_name, "_", year)] %>%
  # === number of observations by state-county code ===#
  .[, num_obs := .N, by = sc_dry] %>%
  # === keep only if more than 10 observations are available ===#
  .[num_obs >= 10, ] %>%
  .[, num_obs := NULL] %>%
  # === order (not necessary) ===#
  .[order(sc_code, year), ] %>%
  dplyr::relocate(sc_code, year)


## -------------------------------------------------------------------
saveRDS(final_data, here("Data/data-processed/final_data.rds"))
