==> Codes/1_Analysis.rmd <==
---
title: "Regression Analysis and Prediction"
author: "Taro Mieno"
---

# Prepare data

```{r}
#--- read data ---#
final_data <-
  here("Data/data-processed/final_data.rds") %>%
  readRDS()

#--- Data used for regression analysis ---#
reg_data <-
  prepare_reg_data(
    data = final_data,
    #--- saturated thickness has to be at least 9 meters ---#
    sat_thld_m = 9,
    #--- at least 75% of total county area has to be overlapped with HPA ---#
    ir_share_thld = 0.75,
    #--- drop obs with extreme balance ---#
    balance_thld = c(-200, 1200),
    #--- number of saturated thickness categories ---#
    sat_cat_num = 3,
    #--- select states of focus ---#
    # no observations are dropped with the current set of states
    state_ls =
      list(
        corn = c("CO", "KS", "NE", "NM", "SD", "TX", "WY"),
        soy = c("KS", "NE", "TX")
      )
  ) %>%
  data.table() %>%
  setnames("crop_type", "crop")
```

# Regression and prediciton (Main)

```{r}
main_analysis <-
  reg_data %>%
  rowwise() %>%
  dplyr::mutate(sc_base = "31011") %>%
  #--- the same sequence used for all the bootstrap iterations ---#
  dplyr::mutate(balance_seq = list(
    seq(
      min(data$balance),
      max(data$balance),
      length = 100
    )
  )) %>%
  dplyr::mutate(sat_seq_eval = list(
    c(10, 40, 70, 100)
  )) %>%
  dplyr::mutate(share_reg_data = list(
    data %>%
      .[, comp := sum(ir == "nir" | ir == "ir"), by = .(sc_code, year)] %>%
      .[comp == 2, ] %>%
      .[, acres_ratio := acres / sum(acres), by = .(sc_code, year)] %>%
      .[ir == "ir" & ir_area_ratio >= 0.75, ] %>%
      .[, balance_avg := mean(balance), by = sc_code] %>%
      .[sat < 180, ]
  )) %>%
  #--- the same sequence used for all the bootstrap iterations ---#
  dplyr::mutate(sat_seq = list(
    share_reg_data[, seq(min(sat),
      max(sat),
      length = 100
    )] %>%
      c(., sat_seq_eval)
  )) %>%
  #--- values of the controls to be used in prediction ---#
  dplyr::mutate(
    sandtotal_med = median(share_reg_data$sandtotal_r),
    silttotal_med = median(share_reg_data$silttotal_r),
    awc_med = median(share_reg_data$awc_r)
  ) %>%
  #++++++++++++++++++++++++++++++++++++
  #+ Main analysis
  #++++++++++++++++++++++++++++++++++++
  dplyr::mutate(yield_analysis_results = list(
    yield_analysis(data, balance_seq, sat_seq_eval, sat_breaks, sat_text_data, sc_base)
  )) %>%
  dplyr::mutate(
    yield_pred_data = list(yield_analysis_results$yield_pred_data),
    yield_semi_res = list(yield_analysis_results$semi_res)
  ) %>%
  dplyr::mutate(yield_dif_test = list(
    lapply(
      2:3,
      \(x) {
        test_dif_in_yield(
          balance_ls = seq(min(data$balance), max(data$balance), length = 50),
          base_q = 1,
          comp_q = x,
          yield_semi_res,
          sat_text_data,
          sc_base
        ) %>%
          .[, comp_q := x]
      }
    ) %>%
      rbindlist()
  )) %>%
  dplyr::mutate(share_pred_data = list(
    share_analysis_gam(
      share_reg_data,
      sat_seq,
      sandtotal_med,
      silttotal_med,
      awc_med
    )
  )) %>%
  dplyr::mutate(avg_yield_pred_data = list(
    get_average_yield(yield_pred_data, share_pred_data)
  ))

saveRDS(main_analysis, "Results/main_analysis.rds")
```

# Bootstrap

```{r}
set.seed(383954)

all_results <-
  main_analysis %>%
  #--- bootsrap data ---#
  dplyr::mutate(boot_results = list(
    parallel::mclapply(
      1:1000,
      \(x) {
        print(x)

        #--- bootstrap data ---#
        boot_data <- boot(data, sc_base)

        #--- yield analysis ---#
        yield_pred_data <-
          yield_analysis(
            boot_data,
            balance_seq,
            sat_seq_eval,
            sat_breaks,
            sat_text_data,
            sc_base,
            boot = TRUE
          )

        #--- share analysis ---#
        share_boot_data <-
          boot_data %>%
          .[, comp := sum(ir == "nir" | ir == "ir"), by = .(sc_code, year)] %>%
          .[comp == 2, ] %>%
          .[, acres_ratio := acres / sum(acres), by = .(sc_code, year)] %>%
          .[ir == "ir" & ir_area_ratio >= 0.75, ] %>%
          .[, balance_avg := mean(balance), by = sc_code] %>%
          .[sat < 180, ]

        share_pred_data <-
          share_analysis_boot(
            ir_share_data = share_boot_data,
            sat_seq,
            sandtotal_med,
            silttotal_med,
            awc_med
          )
        #--- average yield ---#
        avg_yield_pred_data <- get_average_yield(yield_pred_data, share_pred_data)

        return(list(share_pred_data = share_pred_data, avg_yield_pred_data = avg_yield_pred_data))
      },
      mc.cores = parallel::detectCores() - 2
    )
  )) %>%
  dplyr::mutate(
    share_pred_data_boot = list(purrr::map(boot_results, "share_pred_data") %>% rbindlist()),
    avg_yield_pred_data_boot = list(purrr::map(boot_results, "avg_yield_pred_data") %>% rbindlist())
  ) %>%
  dplyr::mutate(share_sum = list(
    share_pred_data_boot[, .(
      ir_share_hat_se = sd(ir_share_hat, na.rm = TRUE),
      dif_ir_share_hat_se = sd(dif_ir_share_hat, na.rm = TRUE)
    ),
    by = .(sat)
    ] %>%
      .[share_pred_data, on = "sat"] %>%
      .[, `:=`(
        lower = ir_share_hat - 1.96 * ir_share_hat_se,
        upper = ir_share_hat + 1.96 * ir_share_hat_se,
        dif_lower = dif_ir_share_hat - 1.96 * dif_ir_share_hat_se,
        dif_upper = dif_ir_share_hat + 1.96 * dif_ir_share_hat_se
      )]
  )) %>%
  dplyr::mutate(avg_yield_sum = list(
    avg_yield_pred_data_boot[, .(
      avg_yield_se = sd(avg_yield, na.rm = TRUE),
      dif_avg_yield_se = sd(dif_avg_yield, na.rm = TRUE)
    ),
    by = .(balance, sat)
    ] %>%
      .[avg_yield_pred_data, on = c("balance", "sat")] %>%
      .[, `:=`(
        upper = avg_yield + 1.96 * avg_yield_se,
        lower = avg_yield - 1.96 * avg_yield_se,
        dif_upper = dif_avg_yield + 1.96 * dif_avg_yield_se,
        dif_lower = dif_avg_yield - 1.96 * dif_avg_yield_se
      )]
  )) %>%
  data.table() %>%
  dplyr::select(
    crop, data, sat_seq, sat_text_data,
    yield_pred_data,
    yield_dif_test,
    share_reg_data,
    share_pred_data,
    share_sum,
    avg_yield_pred_data,
    avg_yield_pred_data_boot,
    avg_yield_sum
  )

saveRDS(all_results, "Results/all_results.rds")
```
